api: fix multispend securityZone type (uint8 not int32)
The InEuint64 struct uses uint8 securityZone, mirroring the existing
/api/v7/spend ABI and the cli-v7 ABIs. Using int32 here produced an
abi-decode mismatch in the contract that surfaced as a bare empty
revert — Solidity can't tell us "your tuple layout is wrong", so the
revert data has no selector and no message.

+++ app/api/v7/multispend/route.ts
@@ -31,7 +31,7 @@ const inEuint64 = {
   type: "tuple",
   components: [
     { name: "ctHash", type: "uint256" },
-    { name: "securityZone", type: "int32" },
+    { name: "securityZone", type: "uint8" },
     { name: "utype", type: "uint8" },
     { name: "signature", type: "bytes" },
   ],
